+++ tests/unit_tests/serialization/test_msgpack.py
@@ -16,9 +16,6 @@
 from base64 import b64encode
 from decimal import Decimal
 
-import msgspec
-import pytest
-
 from nautilus_trader.common.clock import TestClock
 from nautilus_trader.common.enums import ComponentState
 from nautilus_trader.common.factories import OrderFactory
@@ -109,16 +106,6 @@ class TestMsgPackSerializer:
         )
 
         self.serializer = MsgPackSerializer()
-
-    def test_serialize_unknown_object_raises_runtime_error(self):
-        # Arrange, Act
-        with pytest.raises(RuntimeError):
-            self.serializer.serialize({"type": "UNKNOWN"})
-
-    def test_deserialize_unknown_object_raises_runtime_error(self):
-        # Arrange, Act
-        with pytest.raises(RuntimeError):
-            self.serializer.deserialize(msgspec.msgpack.encode({"type": "UNKNOWN"}))
 
     def test_serialize_and_deserialize_fx_instrument(self):
         # Arrange, Act